[MN-34] refactor: Repository를 도메인 단위로 하위 패키지 분리
