+++ src/main/java/com/sprint/mission/sb03monewteam1/batch/job/ArticleCollectJobConfig.java
@@ -3,8 +3,8 @@ package com.sprint.mission.sb03monewteam1.batch.job;
 import com.sprint.mission.sb03monewteam1.entity.Article;
 import com.sprint.mission.sb03monewteam1.entity.InterestKeyword;
 import com.sprint.mission.sb03monewteam1.exception.article.ArticleCollectException;
-import com.sprint.mission.sb03monewteam1.repository.jpa.ArticleRepository;
-import com.sprint.mission.sb03monewteam1.repository.jpa.InterestKeywordRepository;
+import com.sprint.mission.sb03monewteam1.repository.jpa.article.ArticleRepository;
+import com.sprint.mission.sb03monewteam1.repository.jpa.interest.InterestKeywordRepository;
 import com.sprint.mission.sb03monewteam1.service.ArticleService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
