+++ src/main/java/com/sprint/mission/sb03monewteam1/event/listener/NotificationEventListener.java
@@ -8,7 +8,7 @@ import com.sprint.mission.sb03monewteam1.entity.User;
 import com.sprint.mission.sb03monewteam1.event.CommentLikeEvent;
 import com.sprint.mission.sb03monewteam1.event.NewArticleCollectEvent;
 import com.sprint.mission.sb03monewteam1.exception.notification.NotificationSendException;
-import com.sprint.mission.sb03monewteam1.repository.jpa.SubscriptionRepository;
+import com.sprint.mission.sb03monewteam1.repository.jpa.subscription.SubscriptionRepository;
 import com.sprint.mission.sb03monewteam1.service.NotificationService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
