+++ src/main/java/com/sprint/mission/sb03monewteam1/repository/jpa/article/ArticleRepository.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.sb03monewteam1.repository.jpa;
+package com.sprint.mission.sb03monewteam1.repository.jpa.article;
 
 import java.util.Optional;
 import java.util.UUID;
