+++ src/main/java/com/sprint/mission/sb03monewteam1/repository/jpa/articleInterest/ArticleInterestRepository.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.sb03monewteam1.repository.jpa;
+package com.sprint.mission.sb03monewteam1.repository.jpa.articleInterest;
 
 import com.sprint.mission.sb03monewteam1.entity.ArticleInterest;
 import java.util.UUID;
