+++ src/main/java/com/sprint/mission/sb03monewteam1/repository/jpa/articleView/ArticleViewRepository.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.sb03monewteam1.repository.jpa;
+package com.sprint.mission.sb03monewteam1.repository.jpa.articleView;
 
 import com.sprint.mission.sb03monewteam1.entity.ArticleView;
 import java.util.List;
