+++ src/main/java/com/sprint/mission/sb03monewteam1/repository/jpa/comment/CommentRepository.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.sb03monewteam1.repository.jpa;
+package com.sprint.mission.sb03monewteam1.repository.jpa.comment;
 
 import com.sprint.mission.sb03monewteam1.entity.Comment;
 import java.util.List;
