+++ src/main/java/com/sprint/mission/sb03monewteam1/repository/jpa/comment/CommentRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.sb03monewteam1.repository.jpa;
+package com.sprint.mission.sb03monewteam1.repository.jpa.comment;
 
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.Order;
