+++ src/main/java/com/sprint/mission/sb03monewteam1/repository/jpa/interest/InterestKeywordRepository.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.sb03monewteam1.repository.jpa;
+package com.sprint.mission.sb03monewteam1.repository.jpa.interest;
 
 import com.sprint.mission.sb03monewteam1.entity.InterestKeyword;
 import java.util.List;
