+++ src/main/java/com/sprint/mission/sb03monewteam1/repository/jpa/subscription/SubscriptionRepository.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.sb03monewteam1.repository.jpa;
+package com.sprint.mission.sb03monewteam1.repository.jpa.subscription;
 
 import com.sprint.mission.sb03monewteam1.entity.Subscription;
 import java.util.List;
