+++ src/main/java/com/sprint/mission/sb03monewteam1/seeder/SubscriptionDataSeeder.java
@@ -3,9 +3,9 @@ package com.sprint.mission.sb03monewteam1.seeder;
 import com.sprint.mission.sb03monewteam1.entity.Interest;
 import com.sprint.mission.sb03monewteam1.entity.Subscription;
 import com.sprint.mission.sb03monewteam1.entity.User;
-import com.sprint.mission.sb03monewteam1.repository.jpa.SubscriptionRepository;
-import com.sprint.mission.sb03monewteam1.repository.jpa.UserRepository;
-import com.sprint.mission.sb03monewteam1.repository.jpa.InterestRepository;
+import com.sprint.mission.sb03monewteam1.repository.jpa.subscription.SubscriptionRepository;
+import com.sprint.mission.sb03monewteam1.repository.jpa.user.UserRepository;
+import com.sprint.mission.sb03monewteam1.repository.jpa.interest.InterestRepository;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.Set;
